fix: remove artificial 255 cap on ping/pin --count
What changed:
- Widened PingArgs.count (cli.rs) and PingParams.count (ping.rs) from u8 to
  u64, matching the neighboring interval field.
- Added a regression test asserting `-c 1000` parses.
- Documented the fix in CHANGELOG.

Why:
- count was u8, so clap rejected any `-c` above 255 even though the wrapped
  system ping accepts arbitrarily large counts. A user measuring packet loss
  with e.g. `-c 1000` was blocked by a type choice rather than the underlying
  tool — a usability defect in a flag the CLI explicitly advertises.

How it was verified:
- cargo fmt --check, cargo clippy -D warnings, and cargo doc (-D warnings)
  all pass.
- cargo test --all-features --workspace: 36 passed (lib 11 + bin 17 +
  integration 8), 0 failed; the new
  test_cli_parse_ping_accepts_count_above_u8_max passes.

Known risks:
- Public API type change (PingParams.count: u8 -> u64). Acceptable for a 0.x
  crate whose primary consumer is the CLI; widening is source-compatible for
  existing callers passing small integer literals.

+++ src/ping.rs
@@ -7,7 +7,7 @@ use std::process::Command;
 #[derive(Clone, Debug)]
 pub struct PingParams<'a> {
     pub destination: &'a str,
-    pub count: u8,
+    pub count: u64,
     pub size: usize,
     pub ttl: u32,
     pub interval: u64,
